chore: alex - improved login code
